refactor: delete stale codegen; narrow frame_payload_size to uint32
- Remove src/neotape_format_generated.cpp and include/neotape/format_generated.hpp
  (dead code from the old header codegen, replaced by the unified handwritten parser).

- Narrow frame_payload_size from uint64 to uint32 in spec, struct, and parser.
  Max frame payload is (8 MiB - 512) ≈ 8,388,096, well within uint32 range.
  Shifts flags, _reserved, signature, and frame_hash back by 4 bytes each.
  _reserved grows from 246 to 250 bytes to keep the header at 512 bytes total.

- Update tests/test_format.cpp raw-byte offsets and add le32 helper.

- Update docs/spec/01-frame-header.md field table.

+++ include/neotape/format.hpp
@@ -54,7 +54,7 @@ struct FrameHeader {
     uint64_t global_frame_seq_num{0};
     uint64_t logical_slice_seq_num{0};
     uint64_t frame_seq_num_within_channel{0};
-    uint64_t frame_payload_size{0};
+    uint32_t frame_payload_size{0};
     uint64_t flags{0};
     SignatureBytes signature{};
     Hash frame_hash{};
